Minor changes to make build compatible with stampede

+++ src/LLDLA/regLoadStore.cpp
@@ -42,8 +42,8 @@ void LoadToRegs::Prop()
 	cout << "Error: Incorrect dimensions for register load\n";
 	GetInputN(0)->Print();
 	cout << "m_regWidth = " << std::to_string((long long int) m_regWidth) << endl;
-	cout << "*GetInputM(0) != 1 ? " << std::to_string(*GetInputM(0) != 1) << endl;
-	cout << "*GetInputN(0) != m_regWidth ? " << std::to_string(*GetInputN(0) != m_regWidth) << endl;
+	cout << "*GetInputM(0) != 1 ? " << std::to_string((long long int) (*GetInputM(0) != 1)) << endl;
+	cout << "*GetInputN(0) != m_regWidth ? " << std::to_string((long long int) (*GetInputN(0) != m_regWidth)) << endl;
 	throw;
       }
     } else if (*(GetInputN(0)) != 1) {
